feat: Convert email-tracker to Cobra CLI with .env file support
- Convert email-tracker from standalone application to Cobra CLI
- Add shared configuration helpers in internal/config/helpers.go
- Implement .env file loading with proper precedence order:
  1. CLI flags (highest priority)
  2. Environment variables
  3. .env file values
  4. Default values (lowest priority)
- Add CLI flags: --config, --dry-run, --version, --help
- Maintain backward compatibility with existing environment variables
- Follow same Cobra/Fang patterns as main CLI application
- Add comprehensive tests for new configuration functionality
- Update CLAUDE.md with new CLI structure and usage examples

Configuration precedence ensures CLI flags override environment variables,
which override .env file values. The --config flag allows specifying
custom .env file locations for different deployment environments.
