+++ internal/config/config.go
@@ -1,11 +1,9 @@
 package config
 
 import (
-	"bufio"
 	"fmt"
 	"os"
 	"strconv"
-	"strings"
 	"time"
 )
 
@@ -67,7 +65,7 @@ type Config struct {
 // If a .env file exists, it will be loaded first
 func Load() (*Config, error) {
 	// Try to load .env file if it exists
-	loadEnvFile(".env")
+	LoadEnvFile(".env")
 	config := &Config{
 		// Server defaults
 		ServerPort: getEnvOrDefault("SERVER_PORT", "8080"),
@@ -269,86 +267,3 @@ func (c *Config) GetAdminAPIKeyForLogging() string {
 	return c.AdminAPIKey[:4] + "***" + c.AdminAPIKey[len(c.AdminAPIKey)-4:]
 }
 
-// getEnvOrDefault returns environment variable value or default
-func getEnvOrDefault(key, defaultValue string) string {
-	if value := os.Getenv(key); value != "" {
-		return value
-	}
-	return defaultValue
-}
-
-// getEnvDurationOrDefault returns environment variable as duration or default
-func getEnvDurationOrDefault(key string, defaultValue string) time.Duration {
-	if value := os.Getenv(key); value != "" {
-		if duration, err := time.ParseDuration(value); err == nil {
-			return duration
-		}
-	}
-	
-	// Parse default value
-	duration, err := time.ParseDuration(defaultValue)
-	if err != nil {
-		return time.Hour // Fallback to 1 hour
-	}
-	return duration
-}
-
-// getEnvBoolOrDefault returns environment variable as boolean or default
-func getEnvBoolOrDefault(key string, defaultValue bool) bool {
-	if value := os.Getenv(key); value != "" {
-		if parsed, err := strconv.ParseBool(value); err == nil {
-			return parsed
-		}
-	}
-	return defaultValue
-}
-
-// getEnvIntOrDefault returns environment variable as integer or default
-func getEnvIntOrDefault(key string, defaultValue int) int {
-	if value := os.Getenv(key); value != "" {
-		if parsed, err := strconv.Atoi(value); err == nil {
-			return parsed
-		}
-	}
-	return defaultValue
-}
-
-// loadEnvFile loads environment variables from a .env file if it exists
-func loadEnvFile(filename string) {
-	file, err := os.Open(filename)
-	if err != nil {
-		// .env file doesn't exist or can't be opened, which is fine
-		return
-	}
-	defer file.Close()
-
-	scanner := bufio.NewScanner(file)
-	for scanner.Scan() {
-		line := strings.TrimSpace(scanner.Text())
-		
-		// Skip empty lines and comments
-		if line == "" || strings.HasPrefix(line, "#") {
-			continue
-		}
-		
-		// Split on first '=' character
-		parts := strings.SplitN(line, "=", 2)
-		if len(parts) != 2 {
-			continue
-		}
-		
-		key := strings.TrimSpace(parts[0])
-		value := strings.TrimSpace(parts[1])
-		
-		// Remove quotes if present
-		if (strings.HasPrefix(value, "\"") && strings.HasSuffix(value, "\"")) ||
-		   (strings.HasPrefix(value, "'") && strings.HasSuffix(value, "'")) {
-			value = value[1 : len(value)-1]
-		}
-		
-		// Only set if not already set in environment
-		if os.Getenv(key) == "" {
-			os.Setenv(key, value)
-		}
-	}
-}
