+++ internal/config/email_config_cobra_test.go
@@ -1,0 +1,186 @@
+package config
+
+import (
+	"os"
+	"testing"
+)
+
+func TestLoadEmailConfigWithEnvFile(t *testing.T) {
+	// Clean up environment
+	envVars := []string{
+		"GMAIL_CLIENT_ID", "GMAIL_CLIENT_SECRET", "GMAIL_REFRESH_TOKEN",
+		"EMAIL_DRY_RUN", "EMAIL_API_URL", "EMAIL_CHECK_INTERVAL",
+	}
+	for _, v := range envVars {
+		os.Unsetenv(v)
+	}
+	defer func() {
+		for _, v := range envVars {
+			os.Unsetenv(v)
+		}
+	}()
+
+	t.Run("Load from specified env file", func(t *testing.T) {
+		// Create a temporary .env file
+		envContent := `GMAIL_CLIENT_ID=test-client-id
+GMAIL_CLIENT_SECRET=test-client-secret
+EMAIL_DRY_RUN=true
+EMAIL_API_URL=http://test.localhost:8080
+EMAIL_CHECK_INTERVAL=10m
+`
+		
+		// Write to temp file
+		tmpFile, err := os.CreateTemp("", "test*.env")
+		if err != nil {
+			t.Fatal(err)
+		}
+		defer os.Remove(tmpFile.Name())
+		
+		if _, err := tmpFile.WriteString(envContent); err != nil {
+			t.Fatal(err)
+		}
+		tmpFile.Close()
+		
+		// Load config with env file
+		cfg, err := LoadEmailConfigWithEnvFile(tmpFile.Name())
+		if err != nil {
+			t.Fatalf("LoadEmailConfigWithEnvFile failed: %v", err)
+		}
+		
+		// Verify values from .env file were loaded
+		if cfg.Gmail.ClientID != "test-client-id" {
+			t.Errorf("Expected ClientID 'test-client-id', got '%s'", cfg.Gmail.ClientID)
+		}
+		if cfg.Gmail.ClientSecret != "test-client-secret" {
+			t.Errorf("Expected ClientSecret 'test-client-secret', got '%s'", cfg.Gmail.ClientSecret)
+		}
+		if !cfg.Processing.DryRun {
+			t.Errorf("Expected DryRun to be true")
+		}
+		if cfg.API.URL != "http://test.localhost:8080" {
+			t.Errorf("Expected API URL 'http://test.localhost:8080', got '%s'", cfg.API.URL)
+		}
+	})
+
+	t.Run("Environment variables override .env file", func(t *testing.T) {
+		// Create a temporary .env file
+		envContent := `GMAIL_CLIENT_ID=env-file-client-id
+EMAIL_DRY_RUN=false
+`
+		
+		// Write to temp file
+		tmpFile, err := os.CreateTemp("", "test*.env")
+		if err != nil {
+			t.Fatal(err)
+		}
+		defer os.Remove(tmpFile.Name())
+		
+		if _, err := tmpFile.WriteString(envContent); err != nil {
+			t.Fatal(err)
+		}
+		tmpFile.Close()
+		
+		// Set environment variables that should override
+		os.Setenv("GMAIL_CLIENT_ID", "env-var-client-id")
+		os.Setenv("EMAIL_DRY_RUN", "true")
+		defer func() {
+			os.Unsetenv("GMAIL_CLIENT_ID")
+			os.Unsetenv("EMAIL_DRY_RUN")
+		}()
+		
+		// Load config with env file
+		cfg, err := LoadEmailConfigWithEnvFile(tmpFile.Name())
+		if err != nil {
+			t.Fatalf("LoadEmailConfigWithEnvFile failed: %v", err)
+		}
+		
+		// Verify environment variables took precedence
+		if cfg.Gmail.ClientID != "env-var-client-id" {
+			t.Errorf("Expected ClientID 'env-var-client-id' from env var, got '%s'", cfg.Gmail.ClientID)
+		}
+		if !cfg.Processing.DryRun {
+			t.Errorf("Expected DryRun to be true from env var")
+		}
+	})
+
+	t.Run("Empty env file parameter uses default", func(t *testing.T) {
+		// Set minimal required credentials for validation
+		os.Setenv("GMAIL_CLIENT_ID", "test-client")
+		os.Setenv("GMAIL_CLIENT_SECRET", "test-secret")
+		defer func() {
+			os.Unsetenv("GMAIL_CLIENT_ID")
+			os.Unsetenv("GMAIL_CLIENT_SECRET")
+		}()
+		
+		// This should be equivalent to LoadEmailConfig() - just load env variables
+		cfg, err := LoadEmailConfigWithEnvFile("")
+		if err != nil {
+			t.Fatalf("LoadEmailConfigWithEnvFile with empty file failed: %v", err)
+		}
+		
+		// Should use default values since no env vars are set
+		if cfg.API.URL != "http://localhost:8080" {
+			t.Errorf("Expected default API URL, got '%s'", cfg.API.URL)
+		}
+		if cfg.Processing.DryRun != false {
+			t.Errorf("Expected default DryRun false, got %v", cfg.Processing.DryRun)
+		}
+	})
+
+	t.Run("Non-existent env file is ignored", func(t *testing.T) {
+		// Set minimal required credentials for validation
+		os.Setenv("GMAIL_CLIENT_ID", "test-client")
+		os.Setenv("GMAIL_CLIENT_SECRET", "test-secret")
+		defer func() {
+			os.Unsetenv("GMAIL_CLIENT_ID")
+			os.Unsetenv("GMAIL_CLIENT_SECRET")
+		}()
+		
+		// Should not fail, just ignore the missing file
+		cfg, err := LoadEmailConfigWithEnvFile("/nonexistent/path/file.env")
+		if err != nil {
+			t.Fatalf("LoadEmailConfigWithEnvFile with non-existent file failed: %v", err)
+		}
+		
+		// Should use default values
+		if cfg.API.URL != "http://localhost:8080" {
+			t.Errorf("Expected default API URL, got '%s'", cfg.API.URL)
+		}
+	})
+}
+
+func TestLoadEmailConfigBackwardCompatibility(t *testing.T) {
+	// Clean up environment
+	envVars := []string{"GMAIL_CLIENT_ID", "GMAIL_CLIENT_SECRET", "EMAIL_DRY_RUN"}
+	for _, v := range envVars {
+		os.Unsetenv(v)
+	}
+	defer func() {
+		for _, v := range envVars {
+			os.Unsetenv(v)
+		}
+	}()
+
+	// Set some environment variables
+	os.Setenv("GMAIL_CLIENT_ID", "backward-compat-test")
+	os.Setenv("GMAIL_CLIENT_SECRET", "backward-compat-secret")
+	os.Setenv("EMAIL_DRY_RUN", "true")
+	defer func() {
+		os.Unsetenv("GMAIL_CLIENT_ID")
+		os.Unsetenv("GMAIL_CLIENT_SECRET")
+		os.Unsetenv("EMAIL_DRY_RUN")
+	}()
+
+	// Test that LoadEmailConfig() still works as before
+	cfg, err := LoadEmailConfig()
+	if err != nil {
+		t.Fatalf("LoadEmailConfig failed: %v", err)
+	}
+
+	if cfg.Gmail.ClientID != "backward-compat-test" {
+		t.Errorf("Expected ClientID 'backward-compat-test', got '%s'", cfg.Gmail.ClientID)
+	}
+	if !cfg.Processing.DryRun {
+		t.Errorf("Expected DryRun to be true")
+	}
+}
