+++ internal/config/helpers_test.go
@@ -1,0 +1,203 @@
+package config
+
+import (
+	"os"
+	"testing"
+	"time"
+)
+
+func TestGetEnvOrDefault(t *testing.T) {
+	tests := []struct {
+		name     string
+		key      string
+		envValue string
+		expected string
+		default_ string
+	}{
+		{"env var exists", "TEST_VAR", "test_value", "test_value", "default"},
+		{"env var empty", "TEST_VAR", "", "default", "default"},
+		{"env var not set", "NONEXISTENT_VAR", "", "default", "default"},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Clean up environment
+			os.Unsetenv(tt.key)
+			
+			// Set environment variable if specified
+			if tt.envValue != "" {
+				os.Setenv(tt.key, tt.envValue)
+			}
+			
+			result := getEnvOrDefault(tt.key, tt.default_)
+			if result != tt.expected {
+				t.Errorf("getEnvOrDefault(%q, %q) = %q, want %q", tt.key, tt.default_, result, tt.expected)
+			}
+			
+			// Clean up
+			os.Unsetenv(tt.key)
+		})
+	}
+}
+
+func TestGetEnvBoolOrDefaultHelpers(t *testing.T) {
+	tests := []struct {
+		name     string
+		key      string
+		envValue string
+		expected bool
+		default_ bool
+	}{
+		{"true value", "TEST_BOOL_HELPERS", "true", true, false},
+		{"false value", "TEST_BOOL_HELPERS", "false", false, true},
+		{"invalid value", "TEST_BOOL_HELPERS", "invalid", false, false},
+		{"empty value", "TEST_BOOL_HELPERS", "", true, true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Clean up environment
+			os.Unsetenv(tt.key)
+			
+			// Set environment variable if specified
+			if tt.envValue != "" {
+				os.Setenv(tt.key, tt.envValue)
+			}
+			
+			result := getEnvBoolOrDefault(tt.key, tt.default_)
+			if result != tt.expected {
+				t.Errorf("getEnvBoolOrDefault(%q, %v) = %v, want %v", tt.key, tt.default_, result, tt.expected)
+			}
+			
+			// Clean up
+			os.Unsetenv(tt.key)
+		})
+	}
+}
+
+func TestGetEnvIntOrDefaultHelpers(t *testing.T) {
+	tests := []struct {
+		name     string
+		key      string
+		envValue string
+		expected int
+		default_ int
+	}{
+		{"valid int", "TEST_INT_HELPERS", "42", 42, 10},
+		{"invalid int", "TEST_INT_HELPERS", "invalid", 10, 10},
+		{"empty value", "TEST_INT_HELPERS", "", 10, 10},
+		{"negative int", "TEST_INT_HELPERS", "-5", -5, 10},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Clean up environment
+			os.Unsetenv(tt.key)
+			
+			// Set environment variable if specified
+			if tt.envValue != "" {
+				os.Setenv(tt.key, tt.envValue)
+			}
+			
+			result := getEnvIntOrDefault(tt.key, tt.default_)
+			if result != tt.expected {
+				t.Errorf("getEnvIntOrDefault(%q, %v) = %v, want %v", tt.key, tt.default_, result, tt.expected)
+			}
+			
+			// Clean up
+			os.Unsetenv(tt.key)
+		})
+	}
+}
+
+func TestGetEnvDurationOrDefaultHelpers(t *testing.T) {
+	tests := []struct {
+		name     string
+		key      string
+		envValue string
+		expected time.Duration
+		default_ string
+	}{
+		{"valid duration", "TEST_DURATION_HELPERS", "30s", 30 * time.Second, "10s"},
+		{"invalid duration", "TEST_DURATION_HELPERS", "invalid", 10 * time.Second, "10s"},
+		{"empty value", "TEST_DURATION_HELPERS", "", 10 * time.Second, "10s"},
+		{"complex duration", "TEST_DURATION_HELPERS", "1h30m", 90 * time.Minute, "10s"},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			// Clean up environment
+			os.Unsetenv(tt.key)
+			
+			// Set environment variable if specified
+			if tt.envValue != "" {
+				os.Setenv(tt.key, tt.envValue)
+			}
+			
+			result := getEnvDurationOrDefault(tt.key, tt.default_)
+			if result != tt.expected {
+				t.Errorf("getEnvDurationOrDefault(%q, %q) = %v, want %v", tt.key, tt.default_, result, tt.expected)
+			}
+			
+			// Clean up
+			os.Unsetenv(tt.key)
+		})
+	}
+}
+
+func TestLoadEnvFile(t *testing.T) {
+	// Create a temporary .env file
+	envContent := `# Test .env file
+TEST_VAR1=value1
+TEST_VAR2="quoted value"
+TEST_VAR3='single quoted'
+TEST_VAR4=value4
+
+# Another comment
+TEST_VAR5=value5
+`
+	
+	// Write to temp file
+	tmpFile, err := os.CreateTemp("", "test.env")
+	if err != nil {
+		t.Fatal(err)
+	}
+	defer os.Remove(tmpFile.Name())
+	
+	if _, err := tmpFile.WriteString(envContent); err != nil {
+		t.Fatal(err)
+	}
+	tmpFile.Close()
+	
+	// Clean up environment
+	envVars := []string{"TEST_VAR1", "TEST_VAR2", "TEST_VAR3", "TEST_VAR4", "TEST_VAR5"}
+	for _, v := range envVars {
+		os.Unsetenv(v)
+	}
+	
+	// Load the .env file
+	LoadEnvFile(tmpFile.Name())
+	
+	// Test that values were loaded
+	tests := []struct {
+		key      string
+		expected string
+	}{
+		{"TEST_VAR1", "value1"},
+		{"TEST_VAR2", "quoted value"},
+		{"TEST_VAR3", "single quoted"},
+		{"TEST_VAR4", "value4"},
+		{"TEST_VAR5", "value5"},
+	}
+	
+	for _, tt := range tests {
+		if got := os.Getenv(tt.key); got != tt.expected {
+			t.Errorf("LoadEnvFile: %s = %q, want %q", tt.key, got, tt.expected)
+		}
+	}
+	
+	// Clean up
+	for _, v := range envVars {
+		os.Unsetenv(v)
+	}
+}
